+++ app/src/main/java/com/example/recipeforandroid/Network/NetworkManager2.java
@@ -56,7 +56,6 @@ public class NetworkManager2 extends Application{
         JsonObjectRequest request = new JsonObjectRequest(Request.Method.POST, BASE_URL + "api/login", userLogin, new Response.Listener<JSONObject>() {
             @Override
             public void onResponse(JSONObject response) {
-
                 callBack.onSuccess(response);
             }
         }, new Response.ErrorListener() {
@@ -90,7 +89,7 @@ public class NetworkManager2 extends Application{
         queue.add(request);
     }
 
-    public void getUserRecipes(int id, final NetworkCallback callback) {
+    public void getUserRecipes(long id, final NetworkCallback callback) {
         queue = Volley.newRequestQueue(mContext);
         JsonObjectRequest request = new JsonObjectRequest(BASE_URL +"api/" + id + "/recipeList", new Response.Listener<JSONObject>() {
             @Override
